+++ tests/testthat/test-function_helpers.R
@@ -26,3 +26,20 @@ test_that("names of function arguments can be extracted", {
     character()
   )
 })
+
+test_that("Extraction of function body as character works", {
+  test_fun <- function(x) {
+    stopifnot(is.numeric(x))
+    {x + 1}
+  }
+  expect_equal(
+    function_body(test_fun),
+    "stopifnot(is.numeric(x)) { x + 1 }")
+  expect_equal(
+    function_body(test_fun, braces = TRUE),
+    "{ stopifnot(is.numeric(x)) { x + 1 } }")
+  expect_equal(
+    function_body(test_fun, nchar = 20),
+    "stopifnot(is.nume..."
+  )
+})